Adjust MAMBI Salinity standards to deal with single TF stations
Calculating MAMBI scores in a dataset with a single Tidal Freshwater site would cause the PCA part of the script to fail, as it needs multiple values.  The real score is calculated during the TF part of the script, but the data have to pass through the Saline part of the script first.  As such, two placeholder Tidal Freshwater values (good and bad reference) were added to the Saline Standards file.  This allows the PCA to be calculated for the lone tidal freshwater site, discard it, and then allow the rest of script to continue running.