Ensuring all benthic data get queried
Some benthic grabs do not get assigned a grabeventcode of 1, so the filter was changed to only pull grabs designated for benthic infauna via the filter for "Yes" on the benthicinfauna field.  additionally, some benthic samples had more than one grab - either erroneously or on purpose for stations samled with the 4" cores, where 2 "grabs" were combined.  Lastly, some grabs and trawls were collected on the same occupationdate, which caused some benthic data to be duplicated when the occupation table was joined to the abundance table.  to resolve this, a filter was added to only include occuption data designated for benthic grabs.